Bug Fix: Passphrase generator lag
Instead of calling the random word API however many times the length is set to, the call is only made once thanks to the API's number parameter (As seen on line 50).
The new code takes the output of the API which is a string, and then formats the words given into a list which then the for loop uses the words in the new list to create the password.
The effect this bug fix has is a little noticeable on my hardware. Before generating 12 words it took around 2 in a half seconds sometimes 3, now it takes a little less than 1.